requirements.txt
google-generativeai
python-dotenv
jsonschema
graphviz